Repairs made in Payment Services.

+++ EObrazivanjeProjekat/src/main/java/ftn/eObrazovanjeProjekat/EObrazivanjeProjekat/repository/StudentRepository.java
@@ -9,7 +9,7 @@ import ftn.eObrazovanjeProjekat.EObrazivanjeProjekat.model.Student;
 
 public interface StudentRepository extends JpaRepository<Student, Long>{
 	
-	List<Student> findByUser_id(Integer id);
+	List<Student> findByUser_id(Long id);
 	
 	List<Student> findAllStudent(Long idStudent);
 	Student findByIdStudent(Long idStudent);
